refactor: Extract `compile` module
